Tidying up tests into a seperate testing directory and moving mocks into sw-env
