+++ infra/testing/cli-test-helper.js
