+++ infra/testing/e2e-sw-validator.js
