+++ infra/testing/generate-variant-tests.js
