+++ infra/testing/mock-fetch.js
@@ -1,4 +1,4 @@
-global.fetch = async (request) => {
+module.exports = async (request) => {
   const response = new Response('Response from mock-fetch.js', {
     status: 200,
     statusText: 'ok.',
