+++ infra/testing/validator/fake-run-and-compare.js
